requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9 # For PostgreSQL
alembic==1.12.0        # For DB migrations
redis==5.0.1           # For Redis message queue/cache

# Platform Specific (add as needed)
# google-api-python-client==2.105.0
# google-auth-httplib2==0.1.1
# google-auth-oauthlib==1.1.0
# python-whatsapp-sdk==0.4.0 # Or use requests directly